+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [0.2.0](https://github.com/GerritErpenstein/ionic2-custom-icons/compare/0.1.1...0.2.0) (2016-07-25)
+
+- Added support for tabs
+- Shortened import path: (old path still works)
+
+```import {CUSTOM_ICON_DIRECTIVES} from 'ionic2-custom-icons';```
+
 # [0.1.1](https://github.com/GerritErpenstein/ionic2-custom-icons/compare/0.1.0...0.1.1) (2016-06-29)
 
 - Upgraded to Ionic 2.0.0-beta.10